rtf/APP/requirements.txt
aiohttp==3.6.2
paho-mqtt==1.5.0
aiomqtt==0.1.1
xmltodict==0.12.0
pytest==5.3.4
pytest-mock==2.0.0
transitions==0.7.2
async-timeout==3.0.1
pytest-asyncio==0.10.0
pytest-cov==2.8.1